added some resources(images) and fixed a lot of game state progression also made a mockup of the human and its bow object.

+++ PhysicalElement.cpp
@@ -14,12 +14,12 @@ PhysicalElement::~PhysicalElement() {
 
 }
 
-double PhysicalElement::convertPixelToMeter(const int* pixels) {
+double PhysicalElement::convertPixelToMeter(const float* pixels) {
 	return (double)*pixels / (double)100;
 }
 
-int PhysicalElement::convertMeterToPixel(const double* meters) {
-	return (int)round(*meters * 100);
+float PhysicalElement::convertMeterToPixel(const double* meters) {
+	return (float)round(*meters * 100);
 }
 
 void PhysicalElement::setAsCameraFocus() {
